requirements.txt
Django==4.2.0
djangorestframework==3.14.0
# djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
django-cors-headers==4.7.0
Pillow==10.0.0
